# chore: Optimize parial rego execution byte allocations