--- a/coderd/rbac/astvalue.go
+++ b/coderd/rbac/astvalue.go
@@ -134,6 +134,9 @@ func (z Object) regoValue() ast.Value {
 }
 
 func (role Role) regoValue() ast.Value {
+	if role.cachedRegoValue != nil {
+		return *role.cachedRegoValue
+	}
 	orgMap := ast.NewObject()
 	for k, p := range role.Org {
 		orgMap.Insert(ast.StringTerm(k), ast.NewTerm(regoSlice(p)))