Preventing some allocations by reusing the same prepared partial query. Unfortunately the bulk of the work happens when the inputs are passed in, which cannot be done in advance.

The unit tests that are most affected by this change are `Owner` role users. Since the partial execution results in a single empty query. Meaning most of the memory allocations happens before the input is passed in.

# Before

```
cpu: Intel(R) Core(TM) i7-6770HQ CPU @ 2.60GHz
BenchmarkRBACFilter/PrepareOnly-NoRoles-8                     91          15761998 ns/op         2693954 B/op      76514 allocs/op
BenchmarkRBACFilter/PrepareOnly-Admin-8                       76          15272378 ns/op         3154366 B/op      93904 allocs/op
BenchmarkRBACFilter/PrepareOnly-OrgAdmin-8                    61          23254933 ns/op         4741804 B/op     133793 allocs/op
BenchmarkRBACFilter/PrepareOnly-OrgMember-8                   51          21488792 ns/op         4609187 B/op     130032 allocs/op
BenchmarkRBACFilter/PrepareOnly-ManyRoles-8                   46          25034282 ns/op         5267128 B/op     147494 allocs/op
BenchmarkRBACFilter/PrepareOnly-AdminWithScope-8              78          13232737 ns/op         2932466 B/op      85865 allocs/op
```

# After

```
cpu: Intel(R) Core(TM) i7-6770HQ CPU @ 2.60GHz
BenchmarkRBACFilter/PrepareOnly-NoRoles-8                    205           6279153 ns/op         1048482 B/op      29024 allocs/op
BenchmarkRBACFilter/PrepareOnly-Admin-8                      130           8191177 ns/op         1495304 B/op      46161 allocs/op
BenchmarkRBACFilter/PrepareOnly-OrgAdmin-8                    82          13969795 ns/op         3081619 B/op      86025 allocs/op
BenchmarkRBACFilter/PrepareOnly-OrgMember-8                   86          14498713 ns/op         2944946 B/op      82249 allocs/op
BenchmarkRBACFilter/PrepareOnly-ManyRoles-8                   72          25139446 ns/op         3588279 B/op      99513 allocs/op
BenchmarkRBACFilter/PrepareOnly-AdminWithScope-8             150           7135444 ns/op         1276023 B/op      38142 allocs/op
```